Refacto generators module to use alquimia as an adaptor
