+++ fair_forge/generators/__init__.py
@@ -3,11 +3,13 @@
 This module provides synthetic dataset generators for creating test datasets
 from context documents to evaluate AI assistants.
 
+The BaseGenerator class uses LangChain's BaseChatModel interface, allowing
+any LangChain-compatible model to be used for generation.
+
 Available generators:
-- AlquimiaGenerator: Uses Alquimia's internal LLM endpoint
-- OpenAIGenerator: Uses OpenAI's models via LangChain (requires OPENAI_API_KEY)
-- GroqGenerator: Uses Groq's fast inference API via LangChain (requires GROQ_API_KEY)
-- LangChainGenerator: Base class for any LangChain-compatible model
+- BaseGenerator: Base class that accepts any LangChain BaseChatModel
+- AlquimiaGenerator: Implementation using Alquimia's agent API (adapter)
+- AlquimiaChatModel: LangChain adapter for Alquimia agents
 
 Chunk Selection Strategies:
 - SequentialStrategy: Process all chunks in order (default behavior)
@@ -17,9 +19,43 @@ Features:
 - Independent query generation: Generate unrelated questions per chunk
 - Conversation mode: Generate coherent multi-turn conversations where each
   turn builds on the previous one within a chunk
+
+Usage with LangChain models:
+    ```python
+    from langchain_openai import ChatOpenAI
+    from fair_forge.generators import BaseGenerator, create_markdown_loader
+
+    # Create any LangChain-compatible model
+    model = ChatOpenAI(model="gpt-4o-mini")
+
+    # Create generator with the model
+    generator = BaseGenerator(model=model)
+
+    # Generate test datasets
+    loader = create_markdown_loader()
+    datasets = await generator.generate_dataset(
+        context_loader=loader,
+        source="./docs/knowledge_base.md",
+        assistant_id="my-assistant",
+    )
+    ```
+
+Usage with Alquimia:
+    ```python
+    from fair_forge.generators import create_alquimia_generator
+
+    generator = create_alquimia_generator(
+        base_url="https://api.alquimia.ai",
+        api_key="your-key",
+        agent_id="your-agent",
+        channel_id="your-channel",
+    )
+
+    datasets = await generator.generate_dataset(...)
+    ```
 """
 
-from typing import Literal, Optional
+from typing import Optional
 
 from loguru import logger
 
@@ -34,11 +70,8 @@ from fair_forge.schemas.generators import (
     GeneratedQuery,
 )
 
-from .alquimia_generator import AlquimiaGenerator
+from .alquimia_generator import AlquimiaChatModel, AlquimiaGenerator
 from .context_loaders import LocalMarkdownLoader
-from .groq_generator import GroqGenerator
-from .langchain_generator import LangChainGenerator
-from .openai_generator import OpenAIGenerator
 from .strategies import RandomSamplingStrategy, SequentialStrategy
 
 
@@ -48,6 +81,7 @@ def create_alquimia_generator(
     agent_id: str,
     channel_id: str,
     api_version: str = "",
+    use_structured_output: bool = True,
 ) -> AlquimiaGenerator:
     """Create an Alquimia agent-based generator.
 
@@ -61,9 +95,20 @@ def create_alquimia_generator(
         agent_id: Agent/assistant identifier for query generation
         channel_id: Channel identifier
         api_version: API version (optional)
+        use_structured_output: If True, use structured output parsing
 
     Returns:
         AlquimiaGenerator: Configured generator instance
+
+    Example:
+        ```python
+        generator = create_alquimia_generator(
+            base_url="https://api.alquimia.ai",
+            api_key="your-key",
+            agent_id="your-agent",
+            channel_id="your-channel",
+        )
+        ```
     """
     logger.info("Creating Alquimia generator")
     return AlquimiaGenerator(
@@ -72,67 +117,6 @@ def create_alquimia_generator(
         agent_id=agent_id,
         channel_id=channel_id,
         api_version=api_version,
-    )
-
-
-def create_openai_generator(
-    model_name: str = "gpt-4o-mini",
-    temperature: float = 0.7,
-    max_tokens: Optional[int] = 2048,
-    api_key: Optional[str] = None,
-    use_structured_output: bool = True,
-) -> OpenAIGenerator:
-    """Create an OpenAI-based generator.
-
-    The API key is read from the OPENAI_API_KEY environment variable by default.
-
-    Args:
-        model_name: OpenAI model name (default: "gpt-4o-mini")
-        temperature: Sampling temperature (0.0-2.0)
-        max_tokens: Maximum tokens in response
-        api_key: Optional API key (defaults to OPENAI_API_KEY env var)
-        use_structured_output: If True, use with_structured_output() for parsing
-
-    Returns:
-        OpenAIGenerator: Configured generator instance
-    """
-    logger.info(f"Creating OpenAI generator with model: {model_name}")
-    return OpenAIGenerator(
-        model_name=model_name,
-        temperature=temperature,
-        max_tokens=max_tokens,
-        api_key=api_key,
-        use_structured_output=use_structured_output,
-    )
-
-
-def create_groq_generator(
-    model_name: str = "llama-3.1-70b-versatile",
-    temperature: float = 0.7,
-    max_tokens: Optional[int] = 2048,
-    api_key: Optional[str] = None,
-    use_structured_output: bool = True,
-) -> GroqGenerator:
-    """Create a Groq Cloud-based generator.
-
-    The API key is read from the GROQ_API_KEY environment variable by default.
-
-    Args:
-        model_name: Groq model name (default: "llama-3.1-70b-versatile")
-        temperature: Sampling temperature (0.0-2.0)
-        max_tokens: Maximum tokens in response
-        api_key: Optional API key (defaults to GROQ_API_KEY env var)
-        use_structured_output: If True, use with_structured_output() for parsing
-
-    Returns:
-        GroqGenerator: Configured generator instance
-    """
-    logger.info(f"Creating Groq generator with model: {model_name}")
-    return GroqGenerator(
-        model_name=model_name,
-        temperature=temperature,
-        max_tokens=max_tokens,
-        api_key=api_key,
         use_structured_output=use_structured_output,
     )
 
@@ -163,63 +147,14 @@ def create_markdown_loader(
     )
 
 
-def create_generator(
-    backend: Literal["alquimia", "openai", "groq"],
-    **kwargs,
-) -> BaseGenerator:
-    """Factory function to create a generator based on type.
-
-    Args:
-        backend: Generator backend type ("alquimia", "openai", or "groq")
-        **kwargs: Backend-specific configuration
-
-    Returns:
-        BaseGenerator: Configured generator instance
-
-    Raises:
-        ValueError: If backend type is unknown
-
-    Examples:
-        >>> # Create OpenAI generator
-        >>> generator = create_generator(
-        ...     backend="openai",
-        ...     model_name="gpt-4o-mini",
-        ... )
-        >>>
-        >>> # Create Groq generator
-        >>> generator = create_generator(
-        ...     backend="groq",
-        ...     model_name="llama-3.1-70b-versatile",
-        ... )
-        >>>
-        >>> # Create Alquimia generator
-        >>> generator = create_generator(
-        ...     backend="alquimia",
-        ...     base_url="https://api.alquimia.ai",
-        ...     api_key="your-key",
-        ...     model_id="your-model",
-        ... )
-    """
-    if backend == "alquimia":
-        return create_alquimia_generator(**kwargs)
-    elif backend == "openai":
-        return create_openai_generator(**kwargs)
-    elif backend == "groq":
-        return create_groq_generator(**kwargs)
-    raise ValueError(
-        f"Unknown generator backend: {backend}. "
-        f"Valid options: 'alquimia', 'openai', 'groq'"
-    )
-
-
 def create_context_loader(
-    loader_type: Literal["markdown"],
+    loader_type: str,
     **kwargs,
 ) -> BaseContextLoader:
     """Factory function to create a context loader based on type.
 
     Args:
-        loader_type: Context loader type
+        loader_type: Context loader type ("markdown")
         **kwargs: Loader-specific configuration
 
     Returns:
@@ -245,18 +180,13 @@ __all__ = [
     "GeneratedConversationOutput",
     # Generator implementations
     "AlquimiaGenerator",
-    "OpenAIGenerator",
-    "GroqGenerator",
-    "LangChainGenerator",
+    "AlquimiaChatModel",
     "LocalMarkdownLoader",
     # Chunk selection strategies
     "SequentialStrategy",
     "RandomSamplingStrategy",
     # Factory functions
     "create_alquimia_generator",
-    "create_openai_generator",
-    "create_groq_generator",
     "create_markdown_loader",
-    "create_generator",
     "create_context_loader",
 ]
