+++ fair_forge/generators/alquimia_generator.py
@@ -1,35 +1,37 @@
-"""Alquimia AI generator implementation."""
+"""Alquimia AI generator adapter implementation.
 
+This module provides an adapter that wraps the Alquimia client as a LangChain
+BaseChatModel, allowing it to be used with the BaseGenerator interface.
+"""
+
+import asyncio
 import json
 import re
 import uuid
 from datetime import datetime
-from typing import Any, Optional
+from typing import Any, Iterator, List, Optional
 
+from langchain_core.callbacks.manager import CallbackManagerForLLMRun
+from langchain_core.language_models.chat_models import BaseChatModel
+from langchain_core.messages import AIMessage, BaseMessage, HumanMessage, SystemMessage
+from langchain_core.outputs import ChatGeneration, ChatResult
+from langchain_core.runnables import Runnable, RunnableConfig
 from loguru import logger
+from pydantic import Field
 
-from fair_forge.schemas.common import Batch, Dataset
 from fair_forge.schemas.generators import (
-    BaseChunkSelectionStrategy,
-    BaseContextLoader,
     BaseGenerator,
-    Chunk,
-    ConversationTurn,
     GeneratedConversationOutput,
     GeneratedQueriesOutput,
-    GeneratedQuery,
 )
 
-from .strategies import SequentialStrategy
 
+class AlquimiaChatModel(BaseChatModel):
+    """LangChain chat model adapter for Alquimia's agent API.
 
-class AlquimiaGenerator(BaseGenerator):
-    """Generator implementation using Alquimia's agent API.
-
-    This generator uses the Alquimia client to invoke an agent that generates
-    synthetic test queries from context documents. The context, seed examples,
-    and number of queries are passed as extra data kwargs that get injected
-    into the agent's system prompt.
+    This adapter wraps the Alquimia client to implement LangChain's BaseChatModel
+    interface, allowing Alquimia agents to be used with any LangChain-compatible
+    tooling.
 
     Args:
         base_url: Alquimia API base URL
@@ -37,36 +39,115 @@ class AlquimiaGenerator(BaseGenerator):
         agent_id: Agent/assistant identifier for query generation
         channel_id: Channel identifier
         api_version: API version (optional)
+
+    Example:
+        ```python
+        model = AlquimiaChatModel(
+            base_url="https://api.alquimia.ai",
+            api_key="your-api-key",
+            agent_id="your-agent-id",
+            channel_id="your-channel-id",
+        )
+
+        # Use with BaseGenerator
+        generator = BaseGenerator(model=model)
+        ```
     """
 
-    def __init__(
-        self,
-        base_url: str,
-        api_key: str,
-        agent_id: str,
-        channel_id: str,
-        api_version: str = "",
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.base_url = base_url.rstrip("/")
-        self.api_key = api_key
-        self.agent_id = agent_id
-        self.channel_id = channel_id
-        self.api_version = api_version
+    base_url: str = Field(description="Alquimia API base URL")
+    api_key: str = Field(description="Alquimia API key")
+    agent_id: str = Field(description="Agent/assistant identifier")
+    channel_id: str = Field(description="Channel identifier")
+    api_version: str = Field(default="", description="API version")
 
-    async def _invoke_agent(
+    @property
+    def _llm_type(self) -> str:
+        """Return the type of LLM."""
+        return "alquimia"
+
+    @property
+    def _identifying_params(self) -> dict[str, Any]:
+        """Return identifying parameters for this LLM."""
+        return {
+            "base_url": self.base_url,
+            "agent_id": self.agent_id,
+            "channel_id": self.channel_id,
+        }
+
+    def _extract_extra_data_from_messages(
+        self, messages: List[BaseMessage]
+    ) -> dict[str, Any]:
+        """Extract extra data (context, num_queries, etc.) from system message.
+
+        The system prompt contains the context and configuration that should
+        be passed to the Alquimia agent as extra_data kwargs.
+
+        Args:
+            messages: List of messages to extract data from
+
+        Returns:
+            dict: Extra data to pass to the Alquimia agent
+        """
+        extra_data: dict[str, Any] = {}
+
+        for message in messages:
+            if isinstance(message, SystemMessage):
+                content = str(message.content)
+
+                # Extract context from the system prompt
+                context_match = re.search(
+                    r"Context:\s*\n(.*?)(?:\n\n|\nExample|\nGenerate)",
+                    content,
+                    re.DOTALL,
+                )
+                if context_match:
+                    extra_data["context"] = context_match.group(1).strip()
+
+                # Extract num_queries from the system prompt
+                num_queries_match = re.search(
+                    r"Generate exactly (\d+) questions", content
+                )
+                if num_queries_match:
+                    extra_data["num_queries"] = int(num_queries_match.group(1))
+
+                # Extract num_turns for conversation mode
+                num_turns_match = re.search(
+                    r"conversation with exactly (\d+) turns", content
+                )
+                if num_turns_match:
+                    extra_data["num_turns"] = int(num_turns_match.group(1))
+                    extra_data["conversation_mode"] = True
+
+                # Extract seed examples if present
+                seed_match = re.search(
+                    r"Example questions \(match this style\):\s*\n(.*?)\n\n",
+                    content,
+                    re.DOTALL,
+                )
+                if seed_match:
+                    seed_text = seed_match.group(1)
+                    seeds = [
+                        line.strip().lstrip("- ")
+                        for line in seed_text.split("\n")
+                        if line.strip()
+                    ]
+                    if seeds:
+                        extra_data["seed_examples"] = seeds
+
+        return extra_data
+
+    async def _ainvoke_agent(
         self,
         query: str,
         session_id: str,
         extra_data: dict[str, Any],
     ) -> str:
-        """Invoke the Alquimia agent and return the response.
+        """Invoke the Alquimia agent asynchronously and return the response.
 
         Args:
             query: User query to send to the agent
             session_id: Session identifier for conversation context
-            extra_data: Additional kwargs to pass to the agent (injected into system prompt)
+            extra_data: Additional kwargs to pass to the agent
 
         Returns:
             str: Complete response from the agent
@@ -79,12 +160,15 @@ class AlquimiaGenerator(BaseGenerator):
             from alquimia_client import AlquimiaClient
         except ImportError as e:
             raise ImportError(
-                "alquimia-client is required for AlquimiaGenerator. "
+                "alquimia-client is required for AlquimiaChatModel. "
                 "Install it with: uv pip install alquimia-client"
             ) from e
 
+        # Strip trailing slash from base_url
+        clean_base_url = self.base_url.rstrip("/")
+
         async with AlquimiaClient(
-            base_url=self.base_url,
+            base_url=clean_base_url,
             api_key=self.api_key,
             api_version=self.api_version,
         ) as client:
@@ -110,14 +194,163 @@ class AlquimiaGenerator(BaseGenerator):
 
             return response
 
-    def _parse_response(self, response: str) -> GeneratedQueriesOutput:
-        """Parse the agent response to extract generated queries.
+    def _invoke_agent_sync(
+        self,
+        query: str,
+        session_id: str,
+        extra_data: dict[str, Any],
+    ) -> str:
+        """Synchronous wrapper for _ainvoke_agent.
+
+        Args:
+            query: User query to send to the agent
+            session_id: Session identifier
+            extra_data: Additional kwargs
+
+        Returns:
+            str: Complete response from the agent
+        """
+        try:
+            loop = asyncio.get_event_loop()
+            if loop.is_running():
+                # Create a new event loop for this thread
+                import concurrent.futures
+
+                with concurrent.futures.ThreadPoolExecutor() as executor:
+                    future = executor.submit(
+                        asyncio.run,
+                        self._ainvoke_agent(query, session_id, extra_data),
+                    )
+                    return future.result()
+            else:
+                return loop.run_until_complete(
+                    self._ainvoke_agent(query, session_id, extra_data)
+                )
+        except RuntimeError:
+            # No event loop exists, create one
+            return asyncio.run(self._ainvoke_agent(query, session_id, extra_data))
+
+    def _generate(
+        self,
+        messages: List[BaseMessage],
+        stop: Optional[List[str]] = None,
+        run_manager: Optional[CallbackManagerForLLMRun] = None,
+        **kwargs: Any,
+    ) -> ChatResult:
+        """Generate a response from the Alquimia agent.
+
+        Args:
+            messages: List of messages in the conversation
+            stop: Optional list of stop sequences (not used by Alquimia)
+            run_manager: Callback manager for LLM run
+            **kwargs: Additional keyword arguments
+
+        Returns:
+            ChatResult: The generated response
+        """
+        # Extract the user query from the last human message
+        user_query = ""
+        for message in reversed(messages):
+            if isinstance(message, HumanMessage):
+                user_query = str(message.content)
+                break
+
+        # Extract extra data from system message
+        extra_data = self._extract_extra_data_from_messages(messages)
+
+        # Generate a session ID for this invocation
+        session_id = f"gen_{uuid.uuid4().hex[:12]}"
+
+        logger.debug(f"Invoking Alquimia agent with session {session_id}")
+
+        # Call the Alquimia agent
+        response = self._invoke_agent_sync(user_query, session_id, extra_data)
+
+        # Return as ChatResult
+        message = AIMessage(content=response)
+        generation = ChatGeneration(message=message)
+
+        return ChatResult(generations=[generation])
+
+    def with_structured_output(
+        self,
+        schema: type,
+        **kwargs: Any,
+    ) -> "AlquimiaStructuredOutputModel":
+        """Return a model that outputs structured data.
+
+        Args:
+            schema: Pydantic model class for output validation
+            **kwargs: Additional configuration
+
+        Returns:
+            AlquimiaStructuredOutputModel: Model that parses output to schema
+        """
+        return AlquimiaStructuredOutputModel(
+            base_model=self,
+            schema=schema,
+        )
+
+
+class AlquimiaStructuredOutputModel(Runnable):
+    """Wrapper for AlquimiaChatModel that parses output to structured schema.
+
+    This class handles parsing the Alquimia agent's response into the expected
+    Pydantic schema (GeneratedQueriesOutput or GeneratedConversationOutput).
+    Inherits from Runnable to support LangChain's pipe operator.
+    """
+
+    def __init__(
+        self,
+        base_model: AlquimiaChatModel,
+        schema: type,
+    ):
+        """Initialize with base model and output schema.
+
+        Args:
+            base_model: The underlying AlquimiaChatModel
+            schema: Pydantic model class for output validation
+        """
+        self.base_model = base_model
+        self._schema = schema
+
+    def invoke(
+        self,
+        input: Any,
+        config: Optional[RunnableConfig] = None,
+        **kwargs,
+    ) -> Any:
+        """Invoke the model and parse output to schema.
+
+        Args:
+            input: Input value (can be dict or BaseMessage list from prompt)
+            config: Optional runnable config
+            **kwargs: Additional arguments
+
+        Returns:
+            Parsed output matching the schema
+        """
+        # Handle input from ChatPromptTemplate (list of messages)
+        if isinstance(input, list) and len(input) > 0:
+            messages = input
+        elif isinstance(input, dict):
+            messages = input.get("messages", [])
+        else:
+            messages = []
+
+        result = self.base_model._generate(messages)
+        content = result.generations[0].message.content
+
+        return self._parse_response(str(content))
+
+    def _parse_response(self, response: str) -> Any:
+        """Parse the response to the output schema.
 
         Args:
             response: Raw response from the agent
 
         Returns:
-            GeneratedQueriesOutput: Parsed structured output
+            Parsed Pydantic model instance
         """
         # Try to extract JSON from markdown code blocks
         pattern = r"```json\s*(\{.*?\})\s*```"
@@ -130,394 +363,169 @@ class AlquimiaGenerator(BaseGenerator):
             if json_match:
                 json_str = json_match.group(0)
             else:
-                # If no JSON found, try to parse as plain text queries
+                # Fall back to plain text parsing
                 return self._parse_plain_text_response(response)
 
         try:
-            return GeneratedQueriesOutput.model_validate_json(json_str)
+            return self._schema.model_validate_json(json_str)
         except Exception as e:
             logger.warning(f"Failed to parse JSON response: {e}, trying plain text")
             return self._parse_plain_text_response(response)
 
-    def _parse_conversation_response(self, response: str) -> GeneratedConversationOutput:
-        """Parse the agent response to extract conversation turns.
-
-        Args:
-            response: Raw response from the agent
-
-        Returns:
-            GeneratedConversationOutput: Parsed structured output
-        """
-        # Try to extract JSON from markdown code blocks
-        pattern = r"```json\s*(\{.*?\})\s*```"
-        match = re.search(pattern, response, re.DOTALL)
-        if match:
-            json_str = match.group(1)
-        else:
-            # Try to find raw JSON
-            json_match = re.search(r"\{.*\}", response, re.DOTALL)
-            if json_match:
-                json_str = json_match.group(0)
-            else:
-                # If no JSON found, try to parse as plain text
-                return self._parse_plain_text_conversation(response)
-
-        try:
-            return GeneratedConversationOutput.model_validate_json(json_str)
-        except Exception as e:
-            logger.warning(f"Failed to parse JSON conversation: {e}, trying plain text")
-            return self._parse_plain_text_conversation(response)
-
-    def _parse_plain_text_response(self, response: str) -> GeneratedQueriesOutput:
+    def _parse_plain_text_response(self, response: str) -> Any:
         """Parse plain text response when JSON parsing fails.
 
         Args:
             response: Raw response from the agent
 
         Returns:
-            GeneratedQueriesOutput: Parsed output with queries extracted from text
+            Parsed output with data extracted from text
         """
-        queries = []
-        lines = response.strip().split("\n")
-
-        for line in lines:
-            line = line.strip()
-            # Skip empty lines
-            if not line:
-                continue
-            # Check if it looks like a question (ends with ?)
-            if line.endswith("?"):
-                # Remove common prefixes like "1.", "- ", "* ", etc.
-                cleaned = re.sub(r"^[\d]+[\.\)]\s*", "", line)
-                cleaned = re.sub(r"^[-*•]\s*", "", cleaned)
-                if cleaned:
-                    queries.append(GeneratedQuery(query=cleaned))
-
-        if not queries:
-            # If no questions found, treat the whole response as one query
-            queries.append(GeneratedQuery(query=response.strip()))
-
-        return GeneratedQueriesOutput(
-            queries=queries,
-            chunk_summary="Parsed from plain text response",
+        from fair_forge.schemas.generators import (
+            ConversationTurn,
+            GeneratedConversationOutput,
+            GeneratedQueriesOutput,
+            GeneratedQuery,
         )
 
-    def _parse_plain_text_conversation(
-        self, response: str
-    ) -> GeneratedConversationOutput:
-        """Parse plain text response as conversation turns.
+        if self._schema == GeneratedQueriesOutput:
+            queries = []
+            lines = response.strip().split("\n")
 
-        Args:
-            response: Raw response from the agent
+            for line in lines:
+                line = line.strip()
+                if not line:
+                    continue
+                if line.endswith("?"):
+                    cleaned = re.sub(r"^[\d]+[\.\)]\s*", "", line)
+                    cleaned = re.sub(r"^[-*•]\s*", "", cleaned)
+                    if cleaned:
+                        queries.append(GeneratedQuery(query=cleaned))
 
-        Returns:
-            GeneratedConversationOutput: Parsed conversation turns
-        """
-        turns = []
-        lines = response.strip().split("\n")
-        turn_number = 1
+            if not queries:
+                queries.append(GeneratedQuery(query=response.strip()))
 
-        for line in lines:
-            line = line.strip()
-            if not line:
-                continue
-            if line.endswith("?"):
-                cleaned = re.sub(r"^[\d]+[\.\)]\s*", "", line)
-                cleaned = re.sub(r"^[-*•]\s*", "", cleaned)
-                if cleaned:
-                    turns.append(
-                        ConversationTurn(
-                            query=cleaned,
-                            turn_number=turn_number,
+            return GeneratedQueriesOutput(
+                queries=queries,
+                chunk_summary="Parsed from plain text response",
+            )
+
+        elif self._schema == GeneratedConversationOutput:
+            turns = []
+            lines = response.strip().split("\n")
+            turn_number = 1
+
+            for line in lines:
+                line = line.strip()
+                if not line:
+                    continue
+                if line.endswith("?"):
+                    cleaned = re.sub(r"^[\d]+[\.\)]\s*", "", line)
+                    cleaned = re.sub(r"^[-*•]\s*", "", cleaned)
+                    if cleaned:
+                        turns.append(
+                            ConversationTurn(
+                                query=cleaned,
+                                turn_number=turn_number,
+                            )
                         )
-                    )
-                    turn_number += 1
+                        turn_number += 1
 
-        if not turns:
-            turns.append(
-                ConversationTurn(
-                    query=response.strip(),
-                    turn_number=1,
+            if not turns:
+                turns.append(
+                    ConversationTurn(
+                        query=response.strip(),
+                        turn_number=1,
+                    )
                 )
+
+            return GeneratedConversationOutput(
+                turns=turns,
+                conversation_summary="Parsed from plain text response",
+                chunk_summary="Parsed from plain text response",
             )
 
-        return GeneratedConversationOutput(
-            turns=turns,
-            conversation_summary="Parsed from plain text response",
-            chunk_summary="Parsed from plain text response",
-        )
-
-    async def generate_queries(
-        self,
-        chunk: Chunk,
-        num_queries: int = 3,
-        seed_examples: Optional[list[str]] = None,
-        custom_system_prompt: Optional[str] = None,  # Not used with Alquimia
-    ) -> list[GeneratedQuery]:
-        """Generate queries for a single chunk using Alquimia agent.
-
-        The context, seed_examples, and num_queries are passed as extra_data
-        kwargs that get injected into the agent's system prompt.
-
-        Args:
-            chunk: Context chunk to generate queries from
-            num_queries: Number of queries to generate
-            seed_examples: Optional example queries for style guidance
-            custom_system_prompt: Not used with Alquimia (ignored)
-
-        Returns:
-            list[GeneratedQuery]: Generated queries
-        """
-        logger.debug(f"Generating {num_queries} queries for chunk {chunk.chunk_id}")
-
-        # Build extra data that will be injected into the system prompt
-        extra_data: dict[str, Any] = {
-            "context": chunk.content,
-            "num_queries": num_queries,
-        }
-
-        if seed_examples:
-            extra_data["seed_examples"] = seed_examples
-
-        # Create a session ID for this generation
-        session_id = f"gen_{chunk.chunk_id}_{uuid.uuid4().hex[:8]}"
-
-        # Query to send to the agent
-        query = (
-            f"Generate {num_queries} diverse test questions based on the provided context. "
-            f"Return the questions in JSON format with 'queries' array containing objects "
-            f"with 'query', 'difficulty' (easy/medium/hard), and 'query_type' "
-            f"(factual/inferential/comparative/analytical) fields."
-        )
-
-        try:
-            response = await self._invoke_agent(query, session_id, extra_data)
-            output = self._parse_response(response)
-            logger.debug(
-                f"Generated {len(output.queries)} queries for chunk {chunk.chunk_id}"
-            )
-            return output.queries
-        except Exception as e:
-            logger.error(f"Failed to generate queries for chunk {chunk.chunk_id}: {e}")
-            raise
-
-    async def generate_conversation(
-        self,
-        chunk: Chunk,
-        num_turns: int = 3,
-        seed_examples: Optional[list[str]] = None,
-        custom_system_prompt: Optional[str] = None,  # Not used with Alquimia
-    ) -> list[ConversationTurn]:
-        """Generate a coherent multi-turn conversation from a chunk.
-
-        Creates follow-up questions where each turn builds on the previous,
-        maintaining a natural conversation flow exploring the chunk's content.
-
-        Args:
-            chunk: The context chunk to generate conversation from.
-            num_turns: Number of conversation turns to generate.
-            seed_examples: Optional example conversations to guide style.
-            custom_system_prompt: Not used with Alquimia (ignored).
-
-        Returns:
-            list[ConversationTurn]: Ordered conversation turns.
-        """
-        logger.debug(
-            f"Generating {num_turns}-turn conversation for chunk {chunk.chunk_id}"
-        )
-
-        # Build extra data that will be injected into the system prompt
-        extra_data: dict[str, Any] = {
-            "context": chunk.content,
-            "num_turns": num_turns,
-            "conversation_mode": True,
-        }
-
-        if seed_examples:
-            extra_data["seed_examples"] = seed_examples
-
-        # Create a session ID for this generation
-        session_id = f"conv_{chunk.chunk_id}_{uuid.uuid4().hex[:8]}"
-
-        # Query to send to the agent
-        query = (
-            f"Generate a coherent {num_turns}-turn conversation based on the provided context. "
-            f"Each question should naturally follow from and build upon the previous one. "
-            f"Start with a broad question, then drill down into specifics. "
-            f"Return in JSON format with 'turns' array containing objects with 'query', "
-            f"'turn_number', 'difficulty', 'query_type', and 'expected_context' fields."
-        )
-
-        try:
-            response = await self._invoke_agent(query, session_id, extra_data)
-            output = self._parse_conversation_response(response)
-            logger.debug(
-                f"Generated {len(output.turns)} turns for chunk {chunk.chunk_id}"
-            )
-            return output.turns
-        except Exception as e:
-            logger.error(
-                f"Failed to generate conversation for chunk {chunk.chunk_id}: {e}"
-            )
-            raise
-
-    async def generate_dataset(
-        self,
-        context_loader: BaseContextLoader,
-        source: str,
-        assistant_id: str,
-        num_queries_per_chunk: int = 3,
-        language: str = "english",
-        seed_examples: Optional[list[str]] = None,
-        custom_system_prompt: Optional[str] = None,  # Not used with Alquimia
-        selection_strategy: Optional[BaseChunkSelectionStrategy] = None,
-        conversation_mode: bool = False,
-    ) -> list[Dataset]:
-        """Generate datasets from a context source.
-
-        Args:
-            context_loader: Loader to chunk the source document
-            source: Path to the context document
-            assistant_id: ID for the assistant being tested
-            num_queries_per_chunk: Queries/turns per chunk
-            language: Language for queries
-            seed_examples: Example queries for style guidance
-            custom_system_prompt: Not used with Alquimia (ignored)
-            selection_strategy: Strategy for selecting/grouping chunks.
-                Defaults to SequentialStrategy (process all chunks once).
-            conversation_mode: If True, generate coherent multi-turn
-                conversations instead of independent queries.
-
-        Returns:
-            list[Dataset]: Generated datasets (one per chunk group from strategy).
-        """
-        logger.info(f"Loading context from: {source}")
-        chunks = context_loader.load(source)
-        logger.info(f"Loaded {len(chunks)} chunks from source")
-
-        # Use default sequential strategy if none provided
-        strategy = selection_strategy or SequentialStrategy()
-        logger.info(f"Using chunk selection strategy: {strategy}")
-
-        datasets: list[Dataset] = []
-
-        # Process each chunk group from the strategy
-        for chunk_group in strategy.select(chunks):
-            session_id = str(uuid.uuid4())
-            batches: list[Batch] = []
-            full_context = "\n\n".join(chunk.content for chunk in chunk_group)
-
-            if conversation_mode:
-                # Generate coherent conversations per chunk
-                for chunk in chunk_group:
-                    turns = await self.generate_conversation(
-                        chunk=chunk,
-                        num_turns=num_queries_per_chunk,
-                        seed_examples=seed_examples,
-                    )
-
-                    for turn in turns:
-                        qa_id = f"{chunk.chunk_id}_t{turn.turn_number}"
-                        batch = Batch(
-                            qa_id=qa_id,
-                            query=turn.query,
-                            assistant="",
-                            ground_truth_assistant="",
-                            observation=f"Conversation turn {turn.turn_number} from chunk: {chunk.chunk_id}",
-                            agentic=_build_conversation_metadata(chunk, turn),
-                            ground_truth_agentic={},
-                        )
-                        batches.append(batch)
-            else:
-                # Generate independent queries per chunk
-                for chunk in chunk_group:
-                    queries = await self.generate_queries(
-                        chunk=chunk,
-                        num_queries=num_queries_per_chunk,
-                        seed_examples=seed_examples,
-                    )
-
-                    for i, generated_query in enumerate(queries):
-                        qa_id = f"{chunk.chunk_id}_q{i + 1}"
-                        batch = Batch(
-                            qa_id=qa_id,
-                            query=generated_query.query,
-                            assistant="",
-                            ground_truth_assistant="",
-                            observation=f"Generated from chunk: {chunk.chunk_id}",
-                            agentic=_build_agentic_metadata(chunk, generated_query),
-                            ground_truth_agentic={},
-                        )
-                        batches.append(batch)
-
-            logger.info(
-                f"Generated {len(batches)} batches for chunk group "
-                f"({len(chunk_group)} chunks)"
-            )
-
-            dataset = Dataset(
-                session_id=session_id,
-                assistant_id=assistant_id,
-                language=language,
-                context=full_context,
-                conversation=batches,
-            )
-            datasets.append(dataset)
-
-        logger.info(
-            f"Generated {len(datasets)} dataset(s) with "
-            f"{sum(len(d.conversation) for d in datasets)} total batches"
-        )
-
-        return datasets
+        else:
+            raise ValueError(f"Unsupported output schema: {self._schema}")
 
 
-def _build_agentic_metadata(
-    chunk: Chunk, generated_query: GeneratedQuery
-) -> dict[str, Any]:
-    """Build agentic metadata for a query batch.
+class AlquimiaGenerator(BaseGenerator):
+    """Generator implementation using Alquimia's agent API.
+
+    This generator wraps the Alquimia client as a LangChain-compatible model,
+    allowing it to be used with the BaseGenerator interface.
+
+    The context, seed examples, and number of queries are passed as extra
+    data kwargs that get injected into the agent's system prompt.
 
     Args:
-        chunk: The source chunk
-        generated_query: The generated query
+        base_url: Alquimia API base URL
+        api_key: Alquimia API key for authentication
+        agent_id: Agent/assistant identifier for query generation
+        channel_id: Channel identifier
+        api_version: API version (optional)
+        use_structured_output: If True, use structured output parsing
 
-    Returns:
-        dict: Agentic metadata
+    Example:
+        ```python
+        generator = AlquimiaGenerator(
+            base_url="https://api.alquimia.ai",
+            api_key="your-api-key",
+            agent_id="your-agent-id",
+            channel_id="your-channel-id",
+        )
+
+        datasets = await generator.generate_dataset(
+            context_loader=loader,
+            source="./docs/knowledge_base.md",
+            assistant_id="my-assistant",
+        )
+        ```
     """
-    metadata: dict[str, Any] = {"chunk_id": chunk.chunk_id}
-    if generated_query.difficulty:
-        metadata["difficulty"] = generated_query.difficulty
-    if generated_query.query_type:
-        metadata["query_type"] = generated_query.query_type
-    return metadata
+
+    def __init__(
+        self,
+        base_url: str,
+        api_key: str,
+        agent_id: str,
+        channel_id: str,
+        api_version: str = "",
+        use_structured_output: bool = True,
+        **kwargs,
+    ):
+        """Initialize the Alquimia generator.
+
+        Args:
+            base_url: Alquimia API base URL
+            api_key: Alquimia API key
+            agent_id: Agent/assistant identifier
+            channel_id: Channel identifier
+            api_version: API version
+            use_structured_output: If True, use structured output parsing
+            **kwargs: Additional configuration
+        """
+        # Create the Alquimia chat model adapter
+        model = AlquimiaChatModel(
+            base_url=base_url,
+            api_key=api_key,
+            agent_id=agent_id,
+            channel_id=channel_id,
+            api_version=api_version,
+        )
+
+        # Store config for reference
+        self.base_url = base_url.rstrip("/")
+        self.api_key = api_key
+        self.agent_id = agent_id
+        self.channel_id = channel_id
+        self.api_version = api_version
+
+        # Initialize parent with the adapter
+        super().__init__(
+            model=model,
+            use_structured_output=use_structured_output,
+            **kwargs,
+        )
 
 
-def _build_conversation_metadata(
-    chunk: Chunk, turn: ConversationTurn
-) -> dict[str, Any]:
-    """Build agentic metadata for a conversation turn batch.
-
-    Args:
-        chunk: The source chunk
-        turn: The conversation turn
-
-    Returns:
-        dict: Agentic metadata including conversation tracking
-    """
-    metadata: dict[str, Any] = {
-        "chunk_id": chunk.chunk_id,
-        "turn_number": turn.turn_number,
-        "conversation_mode": True,
-    }
-    if turn.turn_number > 1:
-        metadata["builds_on"] = f"{chunk.chunk_id}_t{turn.turn_number - 1}"
-    if turn.difficulty:
-        metadata["difficulty"] = turn.difficulty
-    if turn.query_type:
-        metadata["query_type"] = turn.query_type
-    if turn.expected_context:
-        metadata["expected_context"] = turn.expected_context
-    return metadata
-
-
-__all__ = ["AlquimiaGenerator"]
+__all__ = ["AlquimiaGenerator", "AlquimiaChatModel"]
